+++ scripts/install/ubuntu.sh
@@ -108,11 +108,17 @@ apt install -y libicu-dev
 apt install -y build-essential
 
 apt install -y libcurl4-openssl-dev
+apt install -y libcurl4-nss-dev
 apt install -y curl libcurl4-gnutls-dev
 #https://blog.csdn.net/qq_36228377/article/details/123154344
 # ln -s  /usr/include/x86_64-linux-gnu/curl  /usr/include/curl
 if [ ! -d /usr/include/curl ];then
-    ln -s  /usr/include/x86_64-linux-gnu/curl  /usr/include/curl
+	SYS_ARCH=`arch`
+	if [ -f /usr/include/x86_64-linux-gnu/curl ];then
+		ln -s /usr/include/x86_64-linux-gnu/curl /usr/include/curl
+	else
+		ln -s /usr/include/${SYS_ARCH}-linux-gnu/curl /usr/include/curl
+	fi 
 fi
